requirements.txt
fastapi
pydantic
diffusers
mediapy
google-generativeai
torch
gunicorn